Removed Smarty as a view option
Simplified the View again, everything is now back in a single View class, based on Twig. Smarty has been eliminated comletely for brevity's sake.
